Clean up a few more pubternals
* Make NoopExecutionStrategy public
* Made parameterless ModelBuilder constructor public and use it to create initial Model
* Made NonCapturingLazyInitializer real internal and shared
