+++ src/EFCore/ModelBuilder.cs
@@ -45,17 +45,16 @@ namespace Microsoft.EntityFrameworkCore
         }
 
         /// <summary>
-        ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-        ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-        ///     any release. You should only use it directly in your code with extreme caution and knowing that
-        ///     doing so can result in application failures when updating to a new Entity Framework Core release.
+        ///     <para>
+        ///         Initializes a new instance of the <see cref="ModelBuilder" /> class with no conventions.
+        ///     </para>
+        ///     <para>
+        ///         Warning: conventions are typically needed to build a correct model.
+        ///     </para>
         /// </summary>
-        [EntityFrameworkInternal]
-        public ModelBuilder([NotNull] IMutableModel model)
+        public ModelBuilder()
         {
-            Check.NotNull(model, nameof(model));
-
-            _builder = ((Model)model).Builder;
+            _builder = new Model().Builder;
         }
 
         /// <summary>
