+++ src/EFCore/Storage/Internal/ExecutionStrategyFactory.cs
@@ -43,6 +43,6 @@ namespace Microsoft.EntityFrameworkCore.Storage.Internal
         ///     Creates a new  <see cref="IExecutionStrategy" />.
         /// </summary>
         /// <returns>An instance of <see cref="IExecutionStrategy" />.</returns>
-        public virtual IExecutionStrategy Create() => new NoopExecutionStrategy(Dependencies);
+        public virtual IExecutionStrategy Create() => new NonRetryingExecutionStrategy(Dependencies);
     }
 }
