+++ test/EFCore.SqlServer.FunctionalTests/SqlServerDatabaseCreatorTest.cs
@@ -726,7 +726,7 @@ namespace Microsoft.EntityFrameworkCore
 
             protected override IExecutionStrategy CreateDefaultStrategy(ExecutionStrategyDependencies dependencies)
             {
-                return new NoopExecutionStrategy(dependencies);
+                return new NonRetryingExecutionStrategy(dependencies);
             }
         }
